requirements.txt
streamlit
numpy
pillow
scikit-learn
tqdm
# TensorFlow is optional. If you want the TF model-based embeddings, uncomment below.
# tensorflow
